requirements.txt
# Core App
fastapi==0.110.0
uvicorn==0.29.0

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.2.1
playwright==1.43.0
requests==2.31.0

# NLP / Analysis
openai==1.23.6
spacy==3.7.4
tiktoken==0.6.0

# Utils
python-dotenv==1.0.1
httpx==0.27.0
requests==2.31.0

# Background Tasks
celery==5.3.6
redis==5.0.1

# Testing
pytest==8.1.1
httpx[http2]==0.27.0

# Data Models
pydantic==2.6.4

google-api-python-client

pytest-rich 
pytest-sugar